+++ CHANGELOG.md
@@ -5,6 +5,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.0.3] - 2022-06-26
+
+### Added
+
+- New methods to split `GWADataLoader` objects by chromosome and by samples. 
+The latter should come in handy for splitting the samples for training, validation and testing.
+
+### Changed
+
+- Updated implementation of the shrinkage estimator of LD to align it more closely 
+with the original formulation in Wen and Stephens (2010) and implementations in RSS software.
+- Fixed various bugs and errors in the code.
+- Added proper handling for the slice objects in `plot_ld_matrix`
+
 ## [0.0.2] - 2022-06-15
 
 ### Added